fix: don't try to clean up during dry run
You can end up swallowing the error.

+++ src/release.js
@@ -211,11 +211,13 @@ async function release({
         await originalPush();
       }
     } catch (err) {
-      if (shouldCleanUpAfterFailedPush) {
-        await execa('git', ['tag', '-d', ...tags], { cwd: workspaceCwd });
-      }
+      if (!dryRun) {
+        if (shouldCleanUpAfterFailedPush) {
+          await execa('git', ['tag', '-d', ...tags], { cwd: workspaceCwd });
+        }
 
-      await execa('git', ['reset', '--hard', previousCommit], { cwd: workspaceCwd });
+        await execa('git', ['reset', '--hard', previousCommit], { cwd: workspaceCwd });
+      }
 
       throw err;
     }
